Configured first spring-batch job for admin units

+++ project-db-scripts/src/main/resources/db/changelog/sql_init/2020_01_05_CreateTable_Registration.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS car_info.registration
 (
     "id"                                SERIAL NOT NULL,
     "personType"                        VARCHAR(2),
-    "administrativeTerritorialUnitCode" INT,
+    "administrativeTerritorialUnitCode" BIGINT,
     "registrationDate"                  DATE,
     "vehicleRegistrationNumber"         VARCHAR,
     "vehicle"                           INT    NOT NULL,
